Make chimeric junctions from staggered-splitmappings. Very big changes, completely recoded since we were computing them in the wrong way. These are the main ideas behind the changes:
- Determining the block strandedness for unstranded data. Before we were looking at the consensus splice site (css)
sequence and considering they map in the plus when they have css and in the minus when they have the reverse complementary
of the css. This was totally wrong, if they have the reverse complementary of the css means that what was sequenced was
the reverse complementary of the rna, so the trully strand is the opposite to the mapping strand.

- For stranded data the reads should always split-map with css and not with the reverse complementary sequence of the css.
However, these cases are produced by the mapper so we decided to filter out all this junctions in the step where we
make the junctions.

- Also changed the way the maximum beginnig and end are computed, since it was wrong.