CODE. New script correctNMfield_SAMFromGEM.awk
- Takes as input a SAM file. Check if it was produced from a GEM mapping file with gem-2-sam and if it is the case corrects the edit distance to the reference in the NM:i:X optional field. Needed because ge
m-2-sam considers the skipped regions from the reference (N in the CIGAR string) as mismatches. I think it should not be included since it adds mismatches to reads spanning exon-exon junctions, while this i
s not really a mismatch and another mappers like tophat does not consider them as mismatches.